+++ src/lib/utils.ts
@@ -1521,7 +1521,7 @@ export function setDefaultConfig(initialConfig) {
     interpolatedStopText: 'Estimated time of arrival',
     gtfsToHtmlVersion: version,
     linkStopUrls: false,
-    mapStyleUrl: 'https://tiles.openfreemap.org/styles/liberty',
+    mapStyleUrl: 'https://tiles.openfreemap.org/styles/positron',
     menuType: 'jump',
     noDropoffSymbol: '‡',
     noDropoffText: 'No drop off available',